Implement edit comment modal
